Fixed crashes and memory leaks with killfeed

+++ CarWars/CarWars/Engine/Components/RigidbodyComponents/VehicleComponent.cpp
@@ -481,14 +481,14 @@ void VehicleComponent::TakeDamage(WeaponComponent* damager) {
         for (size_t i = 0; i < Game::gameData.playerCount; ++i) {
             PlayerData& player = Game::players[i];
             Entity* killFeed = EntityManager::FindFirstChild(player.camera->GetGuiRoot(), "KillFeed");
-            std::vector<Entity*> rows = EntityManager::GetChildren(killFeed);
 
             Entity* row = ContentManager::LoadEntity("Menu/KillFeedRow.json", killFeed);
             std::vector<GuiComponent*> guis = row->GetComponents<GuiComponent>();
             GuiComponent* player0Gui = guis[0];
             GuiComponent* player1Gui = guis[1];
             GuiComponent* weaponGui = guis[2];
-            rows.push_back(row);
+            
+            std::vector<Entity*> rows = EntityManager::GetChildren(killFeed);
 
             player1Gui->SetText(me->name);
             const glm::vec2 fontDims = player1Gui->GetFontDimensions();
@@ -516,19 +516,21 @@ void VehicleComponent::TakeDamage(WeaponComponent* damager) {
             constexpr size_t maxCount = 5;
 
             static TTween<float, easing::Quint::easeOut>* tween = nullptr;
-            if (tween) tween->Stop();
+            if (tween) Effects::Instance().DestroyTween(tween);
 
             tween = Effects::Instance().CreateTween<float, easing::Quint::easeOut>(0.f, 1.f, 0.5);
             tween->TakeOwnership();
             tween->SetUpdateCallback([rows, maxCount](float& value) mutable {
-                // Tween in positions
+                if (StateManager::GetState() != GameState_Playing) return;
                 for (int j = 0; j < rows.size(); ++j) {
                     Entity* row = rows[j];
 
+                    // Tween in position
                     float start = 30.f * (static_cast<int>(rows.size()) - 2 - j);
                     float end = 30.f * (static_cast<int>(rows.size()) - 1 - j);
                     GuiHelper::SetGuiYPositions(row, 20.f + glm::mix(start, end, value));
 
+                    // Tween in/out opacity
                     for (GuiComponent* gui : row->GetComponents<GuiComponent>()) {
                         if (rows.size() >= maxCount && j < rows.size() - maxCount) {
                             gui->SetOpacity(1.f - value);
@@ -541,6 +543,7 @@ void VehicleComponent::TakeDamage(WeaponComponent* damager) {
 
             if (rows.size() >= maxCount) {
                 tween->SetFinishedCallback([rows, maxCount](float& value) mutable {
+                    if (StateManager::GetState() != GameState_Playing) return;
                     for (size_t i = 0; i < rows.size() - maxCount; ++i) {
                         EntityManager::DestroyEntity(rows[i]);
                     }
